requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.112.0
h11==0.14.0
idna==3.7
MarkupSafe==2.1.5
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
PyMySQL==1.1.1
pyodbc==5.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.1
reportlab==4.2.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.30.5
Werkzeug==3.0.3
